tests/wpt/web-platform-tests/html/browsers/browsing-the-web/history-traversal/PopStateEvent.html

<!doctype html>
<meta charset=utf-8>
<title>Synthetic popstate events</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function () {
  assert_false('initPopStateEvent' in PopStateEvent.prototype,
               'There should be no PopStateEvent#initPopStateEvent');
}, 'initPopStateEvent');

test(function () {
  var popStateEvent = new PopStateEvent("popstate");
  assert_equals(popStateEvent.state, null, "the PopStateEvent.state");
}, "Initial value of PopStateEvent.state must be null");

test(function () {
  var state = history.state;
  var data;
  window.addEventListener('popstate', function (e) {
    data = e.state;
  });
  window.dispatchEvent(new PopStateEvent('popstate', {
    'state': {testdata:true}
  }));
  assert_true(data.testdata,'state data was corrupted');
  assert_equals(history.state, state, "history.state was NOT set by dispatching the event");
}, 'Dispatching a synthetic PopStateEvent');
</script>
